requirements.txt
# ── Core Backend ────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3

# ── LangGraph Multi-Agent Orchestration ─────────────────────────────────────
langgraph==0.2.60
langchain-core==0.3.28

# ── Ollama LLM Client ────────────────────────────────────────────────────────
httpx==0.27.2

# ── Frontend ─────────────────────────────────────────────────────────────────
streamlit==1.41.1

# ── Utilities ────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
requests==2.32.3
